Finished day 4

+++ events_bulletin/src/eventsdetails/EventDetails.js
@@ -1,9 +1,9 @@
 import './EventDetails.css'
 import React, {useEffect, useState} from 'react'
 import {useParams} from 'react-router-dom'
-import {Link} from 'react-router-dom';
 
 const EventDetails = () => {
+
 
     const { id } = useParams();
     const [eventInfo, setEventInfo] = useState([]);
@@ -13,46 +13,57 @@ const EventDetails = () => {
             .then(res => res.json())
             .then(data => setEventInfo(data[0]))
     },[id])
-
-        if(eventInfo.name) {
-            return (
-                <>
-                    <div id="eventInfoContainer">
-                        <div id="eventInfo">
-                            <p>Event Name: <span>{eventInfo.name}</span></p>
-                            <p>Organizer: <span>{eventInfo.organizer_name}</span></p>
-                            <p>Location: <span>{eventInfo.location}</span></p>
-                            <p>Date: <span>{eventInfo.date.substring(0, 10)}</span></p>
-                            <p>Time: <span>{eventInfo.time}</span></p>
-                            <p>Attendees: <span>{eventInfo.attendees.split(',').length}</span></p>
+console.log(eventInfo)
+    if(eventInfo !== undefined && eventInfo.name) {
+        return (
+            <>
+                <div id="eventInfoContainer">
+                    <div id="eventInfo">
+                        <div id="eventPoints">
+                            <p>Event Name:<br></br><span>{eventInfo.name}</span></p>
+                            <p>Organizer:<br></br><span>{eventInfo.organizer_name}</span></p>
+                            <p>Location:<br></br><span>{eventInfo.location}</span></p>
+                            <p>Date:<br></br><span>{eventInfo.date.substring(0, 10)}</span></p>
+                            <p>Time:<br></br><span>{eventInfo.time}</span></p>
+                            <p>Attendees:<br></br><span>{eventInfo.attendees.split(',').length}</span></p>
                         </div>
 
-                        <div id="eventImage">
-                            <img alt="some image here"/><br/>
-                            <div id="idButtons">
-                                <button>Sign Me Up!</button>
-                                <button>Contact Organizer</button>
-                            </div>
-
-                            <div id="modifyButtons">
-                              <input type="button" value="Edit Event" onClick={() => {
-                                window.location=`http://localhost:3000/events/edit/${id}`
-                              }}></input>
-
-                              <input type="button" value="Delete Event" onClick={() => { //DELETE EVENT
-
-                                fetch(`http://localhost:8081/events/${window.location.href.slice(-1)}`, { method: 'DELETE' })
-                                  .then(() => console.log('Deleted!'))
-                                  .catch((error) => console.error('Error:', error));
-                                }}></input>
-
-                            </div>
+                        <div id="eventPoints">
+                            <p>Event Details:<br></br><span>{eventInfo.details}</span></p>
                         </div>
-
                     </div>
-                </>
-            )
-        }
+
+                    <div id="eventImage">
+                        {console.log(eventInfo.imgPath)}
+
+                        <img src={eventInfo.imgPath} alt="logo here"/><br/>
+
+                        <div id=".">
+                            <button>Sign Me Up!</button>
+                            <button>Contact Organizer</button>
+                        </div>
+
+                        <div id="modifyButtons">
+                            <input type="button" value="Edit Event" onClick={() => {
+                            window.location=`http://localhost:3000/events/edit/${id}`
+                            }}></input>
+
+                            <input type="button" value="Delete Event" onClick={() => { //DELETE EVENT
+
+                            fetch(`http://localhost:8081/events/${window.location.href.slice(-1)}`, { method: 'DELETE' })
+                                .then(() => console.log('Deleted!'))
+                                .catch((error) => console.error('Error:', error));
+                            }}></input>
+
+                        </div>
+                    </div>
+
+                </div>
+            </>
+        )
+    } else {
+        return <h1 style={{marginLeft: '1rem'}}>This page does not exist.</h1>
+    }
 }
 
 export default EventDetails;
